Remove frameworks.datajoint package
