+++ dj_link/__init__.py
@@ -1,5 +1,5 @@
 """A tool for linking two DataJoint tables located on different database servers."""
-from .frameworks.datajoint.link import create_link as link
+from .frameworks.link import create_link as link
 from .schemas import LazySchema
 
 __all__ = ["LazySchema", "link"]
